requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
Pillow==10.0.1
drf-yasg==1.21.7


